# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Documentation",
        "title": "GO Simple Store",
        "version": "1.0.0"
    },
    "host": "apiumkm.kitadigi.com",
    "basePath": "/api/v1/user",
    "paths": {
...
        "/carts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List of Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "List of Cart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number start from zero",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Size per page, default `0`",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field, adding dash (`-`) at the beginning means descending and vice versa",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Select specific fields with comma separated",
                        "name": "fields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "custom filters, see [more details](https://github.com/morkid/paginate#filter-format)",
                        "name": "filters",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Cart",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.Page"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Cart"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Create new Cart",
                "parameters": [
                    {
                        "description": "Cart data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CartRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Cart data",
                        "schema": {
                            "$ref": "#/definitions/model.Cart"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        },
...
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a Transaction by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get a Transaction by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction data",
                        "schema": {
                            "$ref": "#/definitions/model.Checkout"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create Payment Transaction by Transaction id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create Payment Transaction by Transaction id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment Link Response",
                        "schema": {
                            "$ref": "#/definitions/model.PaymentLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/lib.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lib.Page": {
            "type": "object",
            "properties": {
                "first": {
                    "description": "indicate first value",
                    "type": "boolean",
                    "example": true
                },
                "items": {
                    "type": "object"
                },
                "last": {
                    "description": "indicate last value",
                    "type": "boolean",
                    "example": false
                },
                "max_page": {
                    "description": "maximum pages for current schema",
                    "type": "integer",
                    "example": 9
                },
                "page": {
                    "description": "current page, start from zero",
                    "type": "integer",
                    "example": 0
                },
                "size": {
                    "description": "size per page, default `10`",
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "description": "total data without filters",
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "description": "total pages",
                    "type": "integer",
                    "example": 10
                },
                "visible": {
                    "description": "current length",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "lib.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "response message",
                    "type": "string"
                },
                "status": {
                    "description": "http status",
                    "type": "integer"
                }
            }
        },
        "model.Asset": {
            "type": "object",
            "properties": {
                "absolute_path": {
                    "description": "AbsolutePath",
                    "type": "string",
                    "example": "image.png"
                },
                "creator_id": {
                    "description": "creator id",
                    "type": "string",
                    "format": "uuid"
                },
                "description": {
                    "description": "Description",
                    "type": "string",
                    "example": "string"
                },
                "file_path": {
                    "description": "File Path",
                    "type": "string"
                },
                "file_size": {
                    "description": "File Size",
                    "type": "number",
                    "format": "float"
                },
                "filename": {
                    "description": "Filename",
                    "type": "string",
                    "example": "image.png"
                },
                "modifier_id": {
                    "description": "modifier id",
                    "type": "string",
                    "format": "uuid"
                },
                "original_filename": {
                    "description": "OriginalFilename",
                    "type": "string",
                    "example": "image.png"
                },
                "relative_path": {
                    "description": "RelativePath",
                    "type": "string",
                    "example": "image.png"
                },
                "title": {
                    "description": "Title",
                    "type": "string",
                    "example": "Image"
                }
            }
        },
        "model.Brand": {
            "type": "object",
            "required": [
                "brand_code",
                "brand_name"
            ],
            "properties": {
                "brand_code": {
                    "type": "string",
                    "example": "XM"
                },
                "brand_name": {
                    "type": "string",
                    "example": "Iphone"
                }
            }
        },
...
        "model.Cart": {
            "type": "object",
            "properties": {
                "contact": {
                    "$ref": "#/definitions/model.Contact"
                },
                "contact_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "product_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
        "model.CartRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "required": [
                "category_code",
                "category_name"
            ],
            "properties": {
                "category_code": {
                    "type": "string",
                    "example": "HP"
                },
                "category_name": {
                    "type": "string",
                    "example": "Handphone"
                }
            }
        },
...
        "model.Checkout": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contact": {
                    "$ref": "#/definitions/model.Contact"
                },
                "contact_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "product_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "qty": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction": {
                    "$ref": "#/definitions/model.Transaction"
                },
                "transaction_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
        "model.Contact": {
            "type": "object",
            "required": [
                "contact_name"
            ],
            "properties": {
                "address": {
                    "description": "Address",
                    "type": "string",
                    "example": "Jl. Aria Putra No.88, RT.09/RW.01, Sawah Baru, Kec. Ciputat, Kota Tangerang Selatan, Banten 15414"
                },
                "alternate_number": {
                    "description": "Alternate Number",
                    "type": "string",
                    "example": "08123456789"
                },
                "city_id": {
                    "description": "City ID",
                    "type": "string",
                    "format": "uuid"
                },
                "contact_name": {
                    "description": "Contact Name                                                                                // Type",
                    "type": "string",
                    "example": "Walk-in-customers"
                },
                "creator_id": {
                    "description": "creator id",
                    "type": "string",
                    "format": "uuid"
                },
                "email": {
                    "description": "Email",
                    "type": "string",
                    "example": "walk-in-customer@mail.com"
                },
                "mobile": {
                    "description": "Mobile",
                    "type": "string",
                    "example": "08123456789"
                },
                "modifier_id": {
                    "description": "modifier id",
                    "type": "string",
                    "format": "uuid"
                },
                "province_id": {
                    "description": "Province ID",
                    "type": "string",
                    "format": "uuid"
                },
                "subdistrict_id": {
                    "description": "Subdistrict ID",
                    "type": "string",
                    "format": "uuid"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "website": {
                    "description": "Website",
                    "type": "string",
                    "example": "www.walk-in-customer.com"
                },
                "zip_code": {
                    "description": "Zip Code",
                    "type": "string",
                    "example": "15414"
                }
            }
        },
...
        "model.PaymentLinkResponse": {
            "type": "object",
            "properties": {
                "checkout_link": {
                    "type": "string",
                    "example": "https://checkout-staging.xendit.co/latest/65a0eaaa63346762f641c28a"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "required": [
                "description",
                "product_name"
            ],
            "properties": {
                "brand": {
                    "$ref": "#/definitions/model.Brand"
                },
                "brand_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "description": {
                    "type": "string",
                    "example": "Samsung Desc"
                },
                "price": {
                    "type": "number",
                    "example": 10000
                },
                "product_asset": {
                    "$ref": "#/definitions/model.ProductAsset"
                },
                "product_name": {
                    "type": "string",
                    "example": "Samsung"
                },
                "quantity": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "model.ProductAsset": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "product_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "model.Transaction": {
            "type": "object",
            "properties": {
                "checkout": {
                    "$ref": "#/definitions/model.Checkout"
                },
                "contact": {
                    "$ref": "#/definitions/model.Contact"
                },
                "contact_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "invoice_no": {
                    "type": "string",
                    "example": "INV-000000000000072270623"
                },
                "total_amount": {
                    "type": "number",
                    "example": 127000
                },
                "transaction_date": {
                    "type": "string",
                    "format": "date-time"
                },
                "transaction_status": {
                    "type": "string",
                    "example": "pending"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "activated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "asset": {
                    "$ref": "#/definitions/model.Asset"
                },
                "asset_id": {
                    "type": "string"
                },
                "creator_id": {
                    "description": "creator id",
                    "type": "string",
                    "format": "uuid"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "is_activated": {
                    "type": "boolean"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "is_password_system_generated": {
                    "type": "boolean"
                },
                "last_login": {
                    "type": "string",
                    "format": "date-time"
                },
                "mobile": {
                    "type": "string"
                },
                "modifier_id": {
                    "description": "modifier id",
                    "type": "string",
                    "format": "uuid"
                },
                "otp_enabled": {
                    "type": "boolean"
                },
                "otp_expiration": {
                    "type": "string",
                    "format": "date-time"
                },
                "password_expiration": {
                    "type": "string",
                    "format": "date-time"
                },
                "password_last_change": {
                    "type": "string",
                    "format": "date-time"
                },
                "reset_password_expiration": {
                    "type": "string",
                    "format": "date-time"
                },
                "username": {
                    "type": "string"
                },
                "verification_expiration": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TokenKey": {
            "type": "apiKey",
            "name": "x-Token",
            "in": "header"
        }
    }
}